requirements.txt
langchain
langchain-google-genai
google-generativeai
pymupdf 
python-docx
pydantic
pydantic[email]
python-dotenv